Remove default_dashboards recipe
The original reason for modifying the default dashboard was to show more
than 4 starred dashboards on the left hand pane, and so we replaced the
home.json file with a custom one that had the limit removed. Since that
was put in place, two things have happened. First, the default dashboard
shipped by grafana has changed, so the default_dashboards recipe now
reverts the default dashboard to an older version with a broken image
and other changes, instead of just removing the limit. Second, it is now
possible to change the default dashboard from inside grafana itself, so
there is no longer a need to replace the file with chef.

This change simply removes the recipe and associated files that replaced
the home.json default dashboard file, and the default dashboard
installed by the grafana package will be used instead.

Signed-off-by: Mark Harrison <mark@mivok.net>